Snake_and_Ladder game of two palyers
# Snake_and_Ladder_Game Using python

Created a two player snake and ladder game from scrach using python and using some unicodes for snakes and dice to make code or a game more intersting.
Try running the snake_and_ladder.py file from the codebase. 
Enjoy Your Game. ThankYou

This is a simple Python code for playing the popular game snakes and ladders without a UI. It works usng functions of Random library of Python for dice and pre-declared start points and endpoints for both snakes and ladders

![image](https://user-images.githubusercontent.com/59397280/123796549-dca69a00-d902-11eb-961c-bdaac5af0cd2.png)
>>>>>>> 2fea78bff28c947970604d905ac92d7e0ed7f373