Added ATMBANK file
The ATM Program In Python is written in Python programming language, This ATM Software Python Project is a simple console-based system that is very easy to use. Talking about the system ATM Program In Python Using Function, it contains various functions which include Account Statement, Withdrawing, Depositing amount and changing the pin.