Create BMI_Calculator.html
A BMI calculator is a tool that helps to find out BMI easily. By using a BMI calculator, one does not need to calculate manually. He can just put his weight in kilogram and height in meter and click the button “Calculate”/or something of the same task. The calculator will give him his BMI result and as well as comment on his health condition. 

This BMI Calculator is made with HTML, CSS, and javascript.