Want to add Drawing with Turtle module in Python
Description:To utilize the turtle strategies and functionalities, we need to import turtle."turtle" comes loaded with the standard Python bundle and need not be introduced remotely. The guide for executing a turtle program follows 4 stages: Import the turtle module;Make a turtle to control. ;Draw around utilizing the turtle strategies. ;Run turtle.done().