PHOTOGRAPHIC MOISAC WITH PYTHON
How to make Photomosaics? 
Peruse the tile pictures, which will supplant the tiles in the first picture. 
Peruse the objective picture and split it into a M×N framework of tiles. 
For each tile, find the best match from the info pictures. 
Make the final mosaic by masterminding the chose input pictures in a M×N network.

Photographic mosaic, the objective picture isn't downsampled, and the coordinating is finished by contrasting every pixel in the square shape with the comparing pixel from every library picture. The square shape in the objective is then supplanted with the library picture that limits the all out distinction. This requires significantly more calculation than the straightforward kind, however the outcomes can be vastly improved since the pixel-by-pixel coordinating can save the goal of the objective picture.